feat: Use settings in templates
