+++ embedded/public/templates/includes/header.tmpl
@@ -27,8 +27,8 @@
         </header>
         <div class="logo">
             <h1 class="type-title">
-                <span class="terminal-prompt">{{.config.SiteTitle}}</span>
+                <span class="terminal-prompt">{{.settings.Title}}</span>
             </h1>
-            <p class="lighter-text">{{.config.SiteSubtitle}}</p>
+            <p class="lighter-text">{{.settings.Subtitle}}</p>
         </div>
 {{end}}
